backend/requirements.txt
flask
flask-cors
pillow
numpy
onnxruntime
openai>=1.0.0
dotenv